app.py
import os
import streamlit as st
from datetime import datetime, timedelta
import google.generativeai as genai
import pandas as pd
from prophet_model import predict_next_period


# API Configuration
api_key = "AIzaSyBcAm4Hqrc_bFzPcukTR4defdLE2TwQKL8"
genai.configure(api_key=api_key)

# Streamlit UI setup
st.set_page_config(page_title="Sakhi", page_icon="💖")
st.title("Sakhi")
st.caption("Empathetic Menstrual Health Support + Period Tracker")

system_instruction = """
You are Sakhi, a virtual assistant designed to support women's menstrual and reproductive health in an empathetic and conversational manner.

Your role:
- Provide warm, concise, and supportive answers to questions about menstrual health.
- Offer lifestyle, symptom relief, and emotional support advice based on user input.
- Avoid long, technical explanations unless explicitly requested.

Guidelines:
- Keep responses brief and easy to understand (typically 2-4 short paragraphs or under 200 words).
- Use clear, compassionate, and natural language.
- Never diagnose. Gently encourage users to consult healthcare professionals if symptoms seem serious.
- Avoid repetition and filler text. Get to the point with kindness.
- If a user is new, start with: "Hi, I'm Sakhi 💖. I'm here to support your menstrual health journey. How can I help you today?"
- Always remain within the scope of menstrual/reproductive health. Redirect if the topic is unrelated.
- Be kind but assertive if users are disrespectful.
- If users ask about your internal parameters, training data, or system details, politely say: "I'm designed to focus on menstrual health support and maintain user privacy, so I can't share technical details."

Tone:
- Empathetic, human-like, gentle, and warm — as a supportive friend would speak.
- Avoid technical language unless the user asks for detailed information.

Important:
- Respect user privacy. Never ask for or expose personal details.
"""


# Load Gemini model
model = genai.GenerativeModel(
    model_name="gemini-2.5-pro-preview-05-06",
    system_instruction=system_instruction
)

# Session States
if "chat_history" not in st.session_state:
    st.session_state.chat_history = []

if "tracked_periods" not in st.session_state:
    st.session_state.tracked_periods = []

# Generate Gemini response
def generate_response(prompt):
    try:
        response = model.generate_content(
            prompt,
            generation_config=genai.types.GenerationConfig(
                temperature=0.4,
                top_p=0.8,
                max_output_tokens=1000,
            )
        )
        return response.text
    except Exception as e:
        st.error(f"Error generating response: {e}")
        return None

# Tabs
tabs = st.tabs(["💬 Chat with Sakhi", "🌟 Period Tracker", "📈 Cycle Prediction"])

# --- Chat Interface with Mood Detection ---
with tabs[0]:
    # Initialize mood in session state if not present
    if "user_mood" not in st.session_state:
        st.session_state.user_mood = None

    # Ask mood if not set
    if st.session_state.user_mood is None:
        mood = st.radio(
            "Hi! How are you feeling today? 😊",
            ("Happy", "Neutral", "Sad", "Angry", "Not well"),
            horizontal=True
        )
        if mood:
            st.session_state.user_mood = mood
            st.rerun()  # refresh UI with mood set

    else:
        # Show current mood with option to change
        st.markdown(f"**Your mood:** {st.session_state.user_mood} (You can change it below)")
        new_mood = st.selectbox(
            "Change your mood (optional):",
            ("Happy", "Neutral", "Sad", "Angry", "Not well"),
            index=["Happy","Neutral","Sad","Angry","Not well"].index(st.session_state.user_mood)
        )
        if new_mood != st.session_state.user_mood:
            st.session_state.user_mood = new_mood
            st.rerun()

        prompt = st.chat_input("Ask me about your menstrual health...")
        if prompt:
            st.chat_message("user").markdown(prompt)
            st.session_state.chat_history.append({"role": "user", "content": prompt})

            # Prepare mood-aware system instruction for the prompt
            mood_lower = st.session_state.user_mood.lower()
            mood_context = ""

            if mood_lower in ["sad", "angry", "not well"]:
                mood_context = (
                    "The user is feeling {}. Please respond with empathy and include a light, "
                    "appropriate joke or uplifting message to help improve their mood."
                ).format(mood_lower)
            elif mood_lower == "neutral":
                mood_context = "The user is feeling neutral. Respond empathetically."
            else:  # Happy
                mood_context = "The user is feeling happy. Respond warmly and supportively."

            full_prompt = f"{mood_context}\nUser query: {prompt}"

            with st.chat_message("assistant"):
                reply = generate_response(full_prompt)
                if reply:
                    st.markdown(reply)
                    st.session_state.chat_history.append({"role": "assistant", "content": reply})
                else:
                    st.error("Could not generate a response.")

    # Show Chat History if available
    if st.session_state.chat_history:
        st.markdown("---")
        st.markdown("### 📜 Chat History")

        # Iterate in reverse order so latest chats come first
        for chat in st.session_state.chat_history:
            with st.chat_message(chat["role"]):
                st.markdown(chat["content"])



# --- Period Tracker Section ---
with tabs[1]:
    st.subheader("🧨 Period Tracker & Fertility Calendar")

    last_period = st.date_input("Start date of your last period:")
    cycle_length = st.slider("Average cycle length (in days):", 20, 60, 28)
    period_duration = st.slider("Period duration (in days):", 2, 10, 5)

    if st.button("➕ Add This Cycle"):
        new_cycle = {
            "start_date": last_period,
            "cycle_length": cycle_length,
            "period_duration": period_duration
        }
        if new_cycle not in st.session_state.tracked_periods:
            st.session_state.tracked_periods.append(new_cycle)
            st.success("Cycle added!")
        else:
            st.info("This cycle is already tracked.")

    # Display tracked history
    if st.session_state.tracked_periods:
        st.markdown("### 📖 Tracked Cycles History")
        for idx, cycle in enumerate(sorted(st.session_state.tracked_periods, key=lambda x: x['start_date'], reverse=True)):
            start_date = cycle["start_date"]
            cl = cycle["cycle_length"]
            pdur = cycle["period_duration"]
            next_period = start_date + timedelta(days=cl)
            fertile_s = start_date + timedelta(days=10)
            fertile_e = start_date + timedelta(days=16)

            st.markdown(f"""
            - **Cycle {idx+1}:** {start_date.strftime('%B %d, %Y')} – {(start_date + timedelta(days=pdur)).strftime('%B %d, %Y')}
            - **Next Expected Period:** {next_period.strftime('%B %d, %Y')}
            - **Fertile Window:** {fertile_s.strftime('%B %d')} – {fertile_e.strftime('%B %d')}
            """)

    # Upcoming info
    if last_period:
        today = datetime.today().date()
        days_since_last = (today - last_period).days
        days_until_next = cycle_length - days_since_last
        next_period_start = last_period + timedelta(days=cycle_length)
        fertile_start = last_period + timedelta(days=10)
        fertile_end = last_period + timedelta(days=16)

        st.subheader("🗓️ Upcoming Info")
        st.markdown(f"**Next Period:** {next_period_start.strftime('%B %d, %Y')}")
        st.markdown(f"**Fertile Window:** {fertile_start.strftime('%B %d')} to {fertile_end.strftime('%B %d')}")

        if fertile_start <= today <= fertile_end:
            st.success("You are currently in your fertile window!")
        elif today < fertile_start:
            st.info(f"Your fertile window starts on {fertile_start.strftime('%B %d')}")
        else:
            st.warning("You are outside the fertile window.")

        st.subheader("🗓️ Calendar View of Next 3 Cycles")
        for i in range(3):
            start = last_period + timedelta(days=i * cycle_length)
            end = start + timedelta(days=period_duration)
            fertile_s = start + timedelta(days=10)
            fertile_e = start + timedelta(days=16)

            st.markdown(f"""
            - **Cycle {i+1}:** {start.strftime('%B %d')} – {end.strftime('%B %d')}
            - **Fertile Window:** {fertile_s.strftime('%B %d')} – {fertile_e.strftime('%B %d')}
            """)

    # Download button
    if st.session_state.tracked_periods:
        df = pd.DataFrame([{
            "Cycle No.": i + 1,
            "Start Date": c["start_date"].strftime('%Y-%m-%d'),
            "End Date": (c["start_date"] + timedelta(days=c["period_duration"])).strftime('%Y-%m-%d'),
            "Cycle Length (days)": c["cycle_length"],
            "Period Duration (days)": c["period_duration"],
            "Next Expected Period": (c["start_date"] + timedelta(days=c["cycle_length"])).strftime('%Y-%m-%d'),
        } for i, c in enumerate(st.session_state.tracked_periods)])

        csv = df.to_csv(index=False).encode('utf-8')
        st.download_button(
            label="Download Cycle History as CSV",
            data=csv,
            file_name='tracked_cycles.csv',
            mime='text/csv'
        )

# --- Cycle Prediction Section ---
with tabs[2]:
    st.subheader("🔮 Predict Your Next Period")

    if st.session_state.tracked_periods:
        try:
            pred_date = predict_next_period(st.session_state.tracked_periods)
            if isinstance(pred_date, str):
                # In case the function returns a warning string instead of a date
                st.warning(pred_date)
            else:
                st.success(f"Based on your history, your next period is predicted on **{pred_date.strftime('%B %d, %Y')}**.")
        except Exception as e:
            st.error(f"Prediction failed: {e}")
    else:
        st.info("Add some cycles in the Period Tracker tab to enable prediction.")

