requirements.txt
streamlit
pandas
prophet
google-generativeai
python-dotenv